requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
idna==3.2
loguru==0.5.3
peewee==3.14.4
pyTelegramBotAPI==4.1.0
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.7
win32-setctime==1.0.3
